// src/repositories_v2/market/types.ts
export type MarketRepoAddressConfig = {
  queryPackageId: string;
  market: string;
};

export type MarketRepoMetadata = {
  whitelist: {
    lending: ReadonlySet<string>;
    collateral: ReadonlySet<string>;
  };
  poolAddresses: Readonly<Record<string, PoolAddress | undefined>>;
  parseCoinNameFromType: (coinType: string) => string;
  parseCoinType: (coinName: string) => string;
  parseSymbol: (coinName: string) => string;
  parseMarketCoinType: (coinName: string) => string;
  parseMarketCoinName: (coinName: string) => string;
  parseSCoinType: (sCoinName: string) => string | undefined;
  getCoinWrappedType: (coinName: string) => MarketPool['coinWrappedType'];
  getCoinDecimal: (coinName: string) => number | undefined;
  parseAprToApy: (apr: number) => number;
};

export type MarketRepoContext = BaseContext & {
  indexer: IndexerDataSource;
  addresses: MarketRepoAddressConfig;
  metadata: MarketRepoMetadata;
};

export type MarketRepoArgs = BaseRepoArgs & {
  indexer: IndexerDataSource;
  addresses: MarketRepoAddressConfig;
  metadata: MarketRepoMetadata;
};
// ...
export type ParsedMarketPoolData = {
  coinType: string;
  maxBorrowRate: number;
  borrowRate: number;
  borrowRateScale: number;
  borrowIndex: number;
  lastUpdated: number;
  cashAmount: number;
  debtAmount: number;
  marketCoinSupplyAmount: number;
  reserveAmount: number;
  reserveFactor: number;
  borrowWeight: number;
  borrowFee: number;
  baseBorrowRate: number;
  borrowRateOnHighKink: number;
  borrowRateOnMidKink: number;
  highKink: number;
  midKink: number;
  minBorrowAmount: number;
  isIsolated: boolean;
  supplyLimit: number;
  borrowLimit: number;
};

export type CalculatedMarketPoolData = {
  baseBorrowApr: number;
  baseBorrowApy: number;
  borrowAprOnHighKink: number;
  borrowApyOnHighKink: number;
  borrowAprOnMidKink: number;
  borrowApyOnMidKink: number;
  coinDecimal: number;
  conversionRate: number;
  maxBorrowApr: number;
  maxBorrowApy: number;
  borrowApr: number;
  borrowApy: number;
  borrowIndex: number;
  growthInterest: number;
  supplyAmount: number;
  supplyCoin: number;
  borrowAmount: number;
  borrowCoin: number;
  reserveAmount: number;
  reserveCoin: number;
  utilizationRate: number;
  supplyApr: number;
  supplyApy: number;
  isIsolated: boolean;
  maxSupplyCoin: number;
  maxBorrowCoin: number;
};
// ...
export type MarketPool = {
  coinName: string;
  symbol: string;
  coinType: string;
  marketCoinType: string;
  sCoinType: string;
  coinWrappedType: CoinWrappedType;
  coinDecimal: number;
  coinPrice: number;
  maxSupplyCoin: number;
  maxBorrowCoin: number;
  isIsolated: boolean;
} & Required<
  Pick<
    ParsedMarketPoolData,
    | 'highKink'
    | 'midKink'
    | 'reserveFactor'
    | 'borrowWeight'
    | 'borrowFee'
    | 'marketCoinSupplyAmount'
    | 'minBorrowAmount'
  >
> &
  CalculatedMarketPoolData;
// ...
export type MarketReadArgs = {
  coinPrices: CoinPrices;
};